Move app.js and navigation.js from shared to admin.
