+++ src/admin/admin_app.js
@@ -1,9 +1,7 @@
 'use strict'
 
 var React = require('react')
-var { Provider } = require('react-redux')
-var store = require('./redux/store.js')
-var App = require('../shared/app.js')
+var App = require('./app.js')
 
 var AdminApp = React.createClass({
   render () {
@@ -13,6 +11,8 @@ var AdminApp = React.createClass({
         add_details={this.add_details}
         children={this.props.children}
       />
-  )} })
+    )
+  }
+})
 
 module.exports = AdminApp
