+++ src/admin/app.js
